Upgrade reversion to 2.0.8. Read WARNING below!
If your database is large, the migration included in this update of
django-reversion will be slow. For example, on a database with 3.3 million rows
in the `reversion_version` table, the migration takes about 20 minutes. This
can be reduced to around 7 minutes on the same size of database by checking out
commit e822bb1 (from https://github.com/etianen/django-reversion/pull/611).