Allow setting Celery worker count in Django admin
When lots of I/O-bound tasks are enqueued, like REST Services pushes to
slow external servers, it can be helpful to increase the number of
asynchronous worker processes well beyond the number of CPU cores. This
allows that without modifying environment variables or having to restart
server processes for changes to take effect.