feat: Grab user email from OpenID Connect SSO signups
This appears to be a problem with the upstream django-allauth
library. There is a base `DefaultSocialAccountAdapter` class that
populates the fields on the "holder" user model which can be used to
set defaults in the sign-in form. However it only uses "basic" fields
to grab these values. For OpenID connect, the email is not included in
these "basic" fields. Therefore although we have an email in the
values provided from the user info endpoint, it is not used to set the
right default.

This fix uses the ability to extend this `DefaultSocialAccountAdapter`
class to set the user email available.