CI(npm-test): Include node version in cache key for node_modules
"Before I forget", let's include node version in the cache key.

The contents of node_modules after `npm install` may differ if the
Node version (or npm) changes. So to prevent cache misuse, this version
string should be included in the cache key somewhere.

We don't quite need this yet, since we only recently started caching
node_modules, and all CI tests are still on the same version
(v16.15.0, with stock npm 8.5.5).

It's just that we'll want this in place before someone starts
experimenting with moving past 16.15.0, because the cache is reused
across different branches.