Col Resize Perf: Skip re-rendering DraggingStyle between drag start/end
This optimization was the main point of separating the DraggingStyle out
from the ColumnWidths style, but I forgot to apply memo() before.

- With memo(), React skips re-rendering the child component if its own
  props are the same as before.
- Without memo(), React re-renders whenever the parent is re-rendered.

So we skip some JS work (and likely some browser style parsing work)
during the drag event by applying memo() to styles that don't need to
change whenever the columns resize.