Pin exact version of react-router-dom (to 6.14.1)
I'd like to make sure whenever we update this library, we do it as a
focused and intentional endeavor, where someone gets to read the patch
notes, and decide how much manual QA is needed.

Patch notes for react-router:

- https://github.com/remix-run/react-router/releases

That might be true even if we weren't using `unstable_` API's.

An alternative would be to use `~6.14.1` to stay on `6.14.x`, allowing
some `npm` commands to sometimes update patch versions if they exist.

---

Package-lock.json already specifies 6.14.1 prior to this commit, because
it updated while I was resolving the merge conflict (6.14.1 qualifies as
^6.13.0).