For 'your query is too short', show a toast message with no icon.
Users might run into this while they're typing a normal search, so a
Red X seems strong and even a yellow Warning is probably too much. A
green checkmark would be silly.

It'd be nice to detect this with something other than a translated
string comparison, but this will probably work, and yellow warning is
an ok fallback.