Fix "The play() request was interrupted" by putting <audio> in document
Uses a react ref on an attached <audio> element to fix this issue:

"Uncaught (in promise) DOMException: The play() request was interrupted
because the media was removed from the document."

This bug was causing first-time "Play" clicks to be ineffective, so you
would have to click twice to start.

It would also generate a full-page error message in development mode
(and a console error in production). The full-page error was making it
inconvenient to test other issues with audio in the data table.