Fix peerDependency ERR from Reflux in npm >= 8.6.0
    "overrides": {
      "reflux": { "react": "$react"}
    },

This tells npm that Reflux should use our project-level version of React
to resolve its dependency on "react". This override applies even though
our project-level React (16) falls outside of the semver range that
Reflux wants (^15.0.2).

Alternatively, this would work:

    "overrides": {
      "react": "$react"
    },

However, the next time we upgrade React, I'd like to see warnings from
other libraries (if applicable) so we can be sure to upgrade them.

We override this for Reflux because it's not likely to receive updates,
yet it tends to continue to work with newer versions of React without
any problem.

References:

  overrides on npm
  https://docs.npmjs.com/cli/v10/configuring-npm/package-json#overrides

  Helpful comment on the Node breakage thread
  https://github.com/npm/cli/issues/4998#issuecomment-1155261806

The error you'd see when using npm >= 8.6.0 looked like:

  npm ERR! code ERESOLVE
  npm ERR! ERESOLVE could not resolve
  npm ERR!
  npm ERR! While resolving: reflux@6.4.1
  npm ERR! Found: react@16.13.1