Allow GET to /token/ to create a new API token
…if none exists. The old behavior of POST is unchanged for the sake of
API stability.

Something in releases prior to 2.023.37 was creating these tokens
implicitly. It's not happening anymore, meaning that new accounts are
left without API tokens, and the front end has never had any facility to
send anything other than a GET to the token endpoint. To avoid the
hassle of backfilling tokens for accounts that have none, or of
rewriting the front end to send POST whenever GET returns a 404, this
change simply allows a GET request to create a token implicitly if none
exists.