Make async Google translations work only on the…
back end. The front end does not have support for this, and the
Celery-based approach that I fixed here goes against what was later done
in #4002. These fixes serve to provide a proof of concept and remove a
hard-coded bucket name. Full support for asynchronous Google MT will
have to be done in a separate PR.