Fix custom exports of grouped NLP questions
These modifications solve the problem of transcripts and translations
for grouped source questions not appearing when custom exports include
only a subset of fields. However, many questions remain about the
original intent of the code.

Given the comment in the original commit, d013bfe0f5, it's likely
there's just extraneous stuff here that needs to be removed (after
understanding that it's safe to do so):

    this should be pared down to just the atts that are used
    in the frontend and formpack