Tolerate submissions that lack `<formhub><uuid>`
These are accepted by KoboCAT but were breaking the submission API
because the logic to rewrite attachment URLs required `formhub/uuid` to
be present in the JSON submission representation.

This change tries first to read the form UUID from the KoboCAT API
response stored at the asset level (`deployment.backend_response`) and,
if that's somehow missing, only then does it attempt to read the UUID
from the submission.