Upgrade TS, fix error, prompt to use local TS in VS Code
Show the prompt that lets VS Code run the local TypeScript version in
your workspace, so it matches the output of `npx tsc`.

---

The error in formGallery/utils.ts is weird. This StackOverflow snippet
checks if something is an instance of an Object or Array. TS2358 aptly
notices that the right-hand side of

   value instanceof Object || value instanceof Array

will never be reached, since an array is also an instanceof Object.

So this could become

   value instanceof Object

…but to preserve the meaning of the code, I've just switched their
order around.

---

Updated TS while I'm at it, since the code still builds without an issue
and it's nice to keep on top of it.