Update react to v18, allowing semver updates
(Also update webpack-dev-server and autoprefixer while we're here)

This commit does not fully build yet; need:

- Fix emerging TS errors, from types updates
- Make sure react-refresh, etc. works
- Make sure all react-based libraries work
- See about this EBADENGINE >=16.17.0 in Storybook's dependencies
- …and more.

Note the react-table override: v6 -> v7 or v8 is a substantial rewrite
with no upgrade guide. My hope is that, as with reflux, it was forwards-
compatible with react v18, and we can upgrade it later.

A variety of deprecation warnings appear when you do a fresh install
without package-lock.json. Here's how to view them again:

    rm -rf ./node_modules
    npm cache clean -f    # superstitiously, to get resolved/integrity *
    npm install
                                * (https://stackoverflow.com/a/75236714)