Validate 'change email' field
If you click the 'change email' button without filling the email input,
in prod it appears to do nothing; in dev there's a fullscreen error.

We should probably show a toast in case of API error, but meanwhile I
through in some light validation to disable the button until there's at
least something in the text input.