Install eslint-plugin-import for separate type import statements…
e.g., not say 'duplicate import' for this code:

   import React from 'react'
   import type { ReactNode } from 'react'

…which used to work fine, but typescript-eslint's no-duplicate-imports
rule has been removed:

- https://github.com/typescript-eslint/typescript-eslint/commit/47eeea9275bd
- https://typescript-eslint.io/rules/no-duplicate-imports/

We could do more with this plugin.

Only skimmed it, but some of the other features might require some
more config.

- https://github.com/import-js/eslint-plugin-import/blob/main/README.md#resolvers

(Meanwhile, I'll try not to think about how many dev dependencies
we bring in just for this cool rule.)