fix(billing): handle string usage limits TASK-1232 (#5227)
### 📣 Summary
Fixes backend code to handle usage limits that come from Stripe metadata
as strings representing either numbers or "unlimited" usage.

### 👀 Preview steps

1. Create a user and sign them up for a paid subscription
2. Make a project with an audio question and make a submission with an
audio file
3. Navigate to NLP processing for that answer. *Manually* write a short
transcription.
4. On billing-addons-backend branch, try running an automatic
translation of that transcription. You will get a 500 error from the
backend because the string for the usage limit is being compared to an
int.
5. On this branch, you will get a successful transcription.

### 💭 Notes
I would like to have set up more unit testing related to 'unlimited'
usage limits, but the only two usage types that can have unlimited usage
currently (and there is no plan to change this) are submissions and
storage. We do not currently check plan limits for these usage types on
the backend, so there are no code paths for testing the 'unlimited'
change in this PR.