refactor(organizations): add configurable redirect route to org validator TASK-975 (#5262)
### 📣 Summary
Makes the `ValidateOrgPermissions` component require a redirect route so
it can be used outside of the account settings area of the app. Moves
the component file to the root router folder.

### 💭 Notes
I made the redirect route a required prop. Previously redirect was a
boolean that defaulted to true. We don't currently have any use case for
not doing a redirect with this component and requiring it reduces
boolean checks.

---------

Co-authored-by: Leszek Pietrzak <leszek@magicznyleszek.xyz>