refactor(organizations): update admin models for improved clarity TASK-965 (#5256)
### 📣 Summary
Reorganize admin models in the `organization` app for improved structure
and readability.


### 📖 Description
This PR refactors the admin models in the `organization` app to improve
code clarity. The updates simplify the structure of the admin classes
and enhance readability by splitting a large file into smaller, more
manageable ones. These changes also make the codebase easier to extend
and maintain in the future. Existing functionality remains unchanged,
ensuring that current workflows are not affected.