feat(organizations): enable data transfer during batch import via Django Admin  (#5266)
### 📣 Summary
Automate user data transfer to the organization during batch imports
from the Django Admin.


### 📖 Description
This PR introduces functionality to transfer user data to the
organization during batch imports performed via the Django Admin
interface. For each user included in the batch, their projects and
associated data are automatically transferred to the organization as
part of the import process. This enhancement streamlines data
management, ensuring consistency and reducing the need for manual
transfers.