style: presentation
Looks like the checks at the bottom of the PR follows the order defined
in ci.yml.

Place npm-test after the python jobs since the npm-test is a 4-entry
matrix right now and buries the others below the scroll fold

Place pytest on top. it's usually there while you're waiting since it
takes the longest, so this time it doesn't have to move.

Make both files use the same order.

Give "detect changed files" a name to make it easier to find when
you're inspecting results