fix(projectHistoryLogs): rename incorrectly-formatted action TASK-944 (#5320)
### 📣 Summary
Change the action `modify_sharing` to `modify-sharing` in project
history logs for consistency.


### 👷 Description for instance maintainers
Change the AuditAction `modify_sharing` to `modify-sharing`, including a
migration to update existing logs. Also adds the migration for all the
new audit action types, although this migration seems to be a no-op in
SQL.



### 👀 Preview steps

1. ℹ️ have an account and a project
2. Go to Project > Settings > Connect Projects and enable data sharing.
Select a specific question to share.
3. 🔴 [on main] Go to `api/v2/assets/<uid>/history`. Notice the action
for the latest project history log is `modify_sharing`
4. 🟢 [on PR] After running migrations, reload the history endpoint.
Notice the action is now `modify-sharing`
5. [on PR] Select another question to share
6. 🟢 [on PR] Reload the endpoint. The latest log should have
action=`modify-sharing`


### 💭 Notes
I didn't even realize this needed a migration until I ran `./manage
makemigrations audit_log`. It seems to work fine without the migration
but this way there will be no unexpected changes next time we have to
make a real migration.
The migration of existing logs should be pretty light since it's a
fairly rare type of log.