refactor(organization): make organizations endpoint more "RESTful" (#5331)
### 📣 Summary
Refactored the organizations endpoint to adhere more closely to RESTful
principles. Added additional properties to the API response.


### 📖 Description
This PR refactors the organizations endpoint to align better with
RESTful standards. The changes enhance the API's usability and
consistency by including more properties in the response and relying on
url as an (even the primary) identifier for resources.