feat(MembersRoute): add actions dropdown to the table of organization members TASK-987 TASK-990 (#5309)
### 📣 Summary
In Members Table (from Account → Team/Organization → Members) add
actions dropdown with one action. The action allows removing a member or
leaving the organization.

### 📖 Description
The availability of "Remove" action is based on the role of the logged
in user. When trying to remove user (or leave organization) a
confirmation prompt is being shown.

### 👀 Preview steps
1. ℹ️ have multiple different users
2. for one of the users (e.g. "joe"), use
http://kf.kobo.local/admin/organizations/organization/ to add multiple
users into joe's organization
3. For one of the users (e.g. "sue") set the role to "admin"
4. enable "Multi-members override" for joe's organization
5. enable feature flag `mmosEnabled`
6. navigate to `#/account/organization/members`
7. 🟢 notice that a table of organization members displays a list of
users added in step 2
8. 🟢 notice that in the table "joe" (as owner) has an ability to remove
all organization members (excluding themselves) including "sue" who's an
admin

As continuation for above steps:
1. log in as "sue"
2. navigate to `#/account/organization/members`
3. 🟢 notice that in the table "joe" (the owner) has no actions available
4. 🟢 notice that any "member" user in the table has actions available,
and if you click "Remove" a confirmation prompt will appear
5. 🟢 notice that for "sue" the action available is "Leave team"/"Leave
organization", and if you click it, a confirmation prompt will appear -
and it will have different content than confirmation prompt for removing
a member

As continuation for above steps:
1. pick a "member" user and use "Remove" action
2. confirm prompt
3. 🟢 verify that user was removed successfuly

### 💭 Notes
Build atop #5281