fix(organizations): limit useOrganizationQuery refetches TASK-1347 (#5346)
### 🗒️ Checklist

1. [x] run linter locally
2. [x] update all related docs (API, README, inline, etc.), if any
3. [x] draft PR with a title `<type>(<scope>)<!>: <title> TASK-1234`
4. [x] tag PR: at least `frontend` or `backend` unless it's global
5. [x] fill in the template below and delete template comments
6. [x] review thyself: read the diff and repro the preview as written
7. [x] open PR & confirm that CI passes
8. [x] request reviewers, if needed
9. [x] delete this section before merging

### 📣 Summary
Avoid unnecessary fetches of organization data


### 📖 Description
After patching changes to organization members react-query was making
two (sometimes one) refetches of backend data for
`useOrganizationQuery`. This was undesired behavior. We want the
organization members list to be refetched, but not the organization
detail itself. To resolve this, I have set the `staleTime` option on
`useOrganizationQuery` to 2 minutes. Since the data will not be
considered stale until 2 minutes passes since the last refresh, these
unnecessary API calls will not be made. Additionally, we have added an
optional param to the Organization query to invalidate cached data on
mount. This is used by the settings page. Since this page is intended
for editing organization data, I think it is preferable to always fetch
the latest data when opening it.


### 👀 Preview steps
1. Sign in as an MMO admin or owner for an org with a few members
2. On the organization members table, change the role of one of the
other members
3. On main, observe the organization data being fetched after a
successful member patch request
4. On this branch, observe no org data fetch
5. On this branch, navigate to the Organization settings page and see
that the organization data is refetched.

### 💭 Notes
I set a very low stale time for the organization. There had been some
discussion about setting a long staleTime, but my thinking here is that
our real goal with this change is to prevent the frontend from spamming
the backend with requests for organization data on every member
mutation. A low staleTime achieves this while still allowing for fairly
frequent automatic updates of the data.

---------

Co-authored-by: Paulo Amorim <pauloamorimbr@gmail.com>