fix(billing): adjust usage page loading spinner logic TASK-1321 (#5323)
### 📣 Summary
Fix error that was causing infinite loading spinner on usage page for
non-Stripe instances.

### 📖 Description
The usage page was expecting addons context to complete loading before
switching from loading spinner to displaying page contents. However, the
usage page needs to be usable on non-Stripe instances addons context
never loads on such instances. I have adjusted the logic for checking
whether or not the page is ready to display accordingly. At the same
time, I also removed the memoization for this logic, as it was
unnecessary and only used in one place.

### 👀 Preview steps

1. Access the usage page with Stripe disabled
2. On main, you will get an infinite loading spinner
3. On this branch, you will see the usage page as expected

Co-authored-by: Anji Tong <ollejna@gmail.com>