feat(projectHistoryLogs): add action label and review all action items and texts TASK-1371 (#5354)
### 📣 Summary
This PR adds labels to all action items and display the new labels in
the filter selection.
Also, all the existing items and texts were reviewed and updated based
on [the actions
list](https://www.notion.so/kobotoolbox/Write-code-for-rendering-all-possible-activity-messages-in-the-UI-1287e515f65480868bbcdc19c085920f?pvs=4#1287e515f654808a83cce66d13abde6c)


### 📖 Description
- A new label property was added to all items in the
`AUDIT_ACTION_TYPES`
- This new property is now being used as the label for the filter
selection items
- The filter selection box was enlarged from 150px to 280px (initially
suggested new size was of 250px, but it was needed to make it even
larger to fit all the existing labels)
- All the items from `AuditActions` enum and all the texts from
`AUDIT_ACTION_TYPES` were reviewed and updated based on [the actions
list](https://www.notion.so/kobotoolbox/Write-code-for-rendering-all-possible-activity-messages-in-the-UI-1287e515f65480868bbcdc19c085920f?pvs=4#1287e515f654808a83cce66d13abde6c)
- Transferring a project and modifying permissions are special cases
where more than 1 username is used in the message. A fix for that with
the current metadata was added.

### 👀 Preview steps
1. ℹ️ have an account and a project
2. Have the activity logs feature flag on with:
`?ff_activityLogsEnabled=true`
3. Transfer a project to another user
4. Navigate to the project's Settings > Activity
5. 🟢 The transfer should be displayed in the activity log
6. 🟢 The destination username should be properly displayed
7. Change the access permission of an user
8. 🟢 The permission change should be displayed in the activity log
9. 🟢 The username who had the permission changes should be displayed in
the log
10. All other logged actions should be displayed in the list with the
proper text