fix(NLP): NLP features broken on non-Stripe instances TASK-1322 (#5337)
Fix bug in NLP feature when stripe is disabled

### 📖 Description
This PR fixes a bug found on the NLP feature that occurred when stripe
was disabled. Some code was refactored and a new unit tests was added to
ensure that no issue arises when stripe is disabled.

### 💭 Notes
Ensure that all the necessary env variables are set up. While testing
this on the local instance, I found that the variable GS_BUCKET_NAME was
not set.

Signed-off-by: Guillermo <lgar89@gmail.com>