feat(auditLogs): add endpoint for retrieving actions TASK-1365 (#5383)
### 📣 Summary
Add a new endpoint for viewing all audit actions taken for an asset.


### 💭 Notes
Added as an `action` so it was easily nested in
`api/v2/asset/<uid>/history`. Confirmed that this does not result in any
unnecessary queries.


Feature/no-change template:
1. ℹ️ have an account 
2. Create a new project. Just give it a name (no other settings) and
some questions in the form builder. Don't deploy it yet.
3. Go to `api/v2/assets/<project_uid>/history/actions`
5. 🟢 The response should be
```
{ "actions": [
    "update-content"
]}
```
6. Change the name
7. Deploy the project
7. 🟢 Reload the endpoint. The response should now be (order doesn't
matter):
```
{ "actions": [
    "update-name",
    "update-content",
    "deploy",
]}
```