fix(BugFixes) Fix error handling for invalid XLS forms TASK-1353 (#5403)
Fix the error handling for invalid survey JSON content when deploying

This PR adds error handling for the case when Pyxform fails when
attempting to deploy a survey with invalid content. A unit test was
included to ensure this works ok.

1. Create a new survey using a bad xls form (see the notion page for a
test file with bad format)
2. Attempt to deploy the form
3. The frontend should show a validation error with a hint of how to fix
it
4. Edit and save the form (labels are added automatically when you enter
the form editor, so there's no need to do anything in this case, just
edit and hit the save button)
5. Attempt to deploy again
6. It should be deployed without issues